+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## master
+## Release 3.5.0
 
 - SERVER: Updated list of latest ASN allocations of 16-bit & 32-bit ASN's from IANA (GH-293). [Thanks @itsbalamurali]